changed layout of schusszettel module to be the project best practice. Removed redundant calls and false cross referencing. tables of schusszettel do not use tablet sessions. They get their token at qr scan and thats it.
