+++ bogenliga/src/app/modules/schusszettel/components/maske1registrierung/register-rueckennummer.component.ts
@@ -1,6 +1,7 @@
 import {Component, Input} from '@angular/core';
 import {Router} from '@angular/router';
-import {SchusszettelService} from '../schusszettel.service';
+import {SchusszettelService} from '@schusszettel/services/schusszettel.service';
+
 
 @Component({
   selector:    'bla-maske1registrierung',
