+++ bogenliga/src/app/modules/schusszettel/components/maske3aktualisierung/warte-bestaetigung.component.ts
@@ -1,6 +1,6 @@
 import {Component, Input} from '@angular/core';
 import {Router} from '@angular/router';
-import {SchusszettelService} from '../schusszettel.service';
+import {SchusszettelService} from '@schusszettel/services/schusszettel.service';
 
 @Component({
   selector:    'bla-maske3aktualisierung',
