+++ bogenliga/src/app/modules/schusszettel/components/maske4wettkampfbeendet/wettkampfbeendet.component.ts
