+++ bogenliga/src/app/modules/schusszettel/components/maske5not-allowed/not-allowed.component.ts
