+++ bogenliga/src/app/modules/schusszettel/components/setup/schusszettel-tablet-admin.component.ts
