+++ bogenliga/src/app/modules/schusszettel/components/tablet/tablet.component.ts
