+++ bogenliga/src/app/modules/schusszettel/schusszettel.routing.ts
@@ -1,7 +1,7 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
-import {SchusszettelTabletAdminComponent} from './setup/schusszettel-tablet-admin.component';
-import {TabletComponent} from './tablet/tablet.component';
+import {SchusszettelTabletAdminComponent} from './components/setup/schusszettel-tablet-admin.component';
+import {TabletComponent} from './components/tablet/tablet.component';
 
 const routes: Routes = [
   {path: 'wettkampfleiter-tablet-view/:id', component: SchusszettelTabletAdminComponent},
